# api/openapi.yaml
openapi: 3.1.0
info:
  title: Origami FarmOS API
  description: Offline-first, tablet-first farm operating system API for Origami Farms.
    See CONSTITUTION.md and product/MVP_SCOPE.md in the repository root for the product
    principles this API enforces (evidence-based recommendations, worker-observes/manager-decides
    separation, event-sourced history).
  version: 0.1.0
paths:
...
  /api/v1/fields:
    post:
      tags:
      - agriculture
      summary: Create Field
      operationId: create_field_api_v1_fields_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FieldCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FieldOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
  /api/v1/harvest:
    post:
      tags:
      - agriculture
      summary: Record Daily Harvest
      description: 'Records the day''s pick and moves the sellable part into real

        inventory, so "168 kg ready for sale" on the Produce screen is stock

        that exists rather than a number someone typed.'
      operationId: record_daily_harvest_api_v1_harvest_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DailyHarvestCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailyHarvestOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
  /api/v1/observations:
    get:
      tags:
      - observations
      summary: List Observations
      operationId: list_observations_api_v1_observations_get
      security:
      - HTTPBearer: []
      parameters:
      - name: farm_id
        in: query
        required: true
        schema:
          type: string
          title: Farm Id
      - name: entity_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Entity Type
      - name: entity_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Entity Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ObservationOut'
                title: Response List Observations Api V1 Observations Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - observations
      summary: Create Observation
      description: 'Constitution: "Workers record observations. Workers do not

        diagnose." [ObservationCreate] structurally has no diagnosis field, so

        this endpoint cannot be used to smuggle one in regardless of caller

        role.'
      operationId: create_observation_api_v1_observations_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ObservationCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObservationOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/production/fields:
    get:
      tags:
      - production
      summary: List Fields
      operationId: list_fields_api_v1_production_fields_get
      security:
      - HTTPBearer: []
      parameters:
      - name: farm_id
        in: query
        required: true
        schema:
          type: string
          title: Farm Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FieldOut'
                title: Response List Fields Api V1 Production Fields Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/production/milk:
    get:
      tags:
      - production
      summary: List Milk Records
      operationId: list_milk_records_api_v1_production_milk_get
      security:
      - HTTPBearer: []
      parameters:
      - name: farm_id
        in: query
        required: true
        schema:
          type: string
          title: Farm Id
      - name: days
        in: query
        required: false
        schema:
          type: integer
          default: 30
          title: Days
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MilkRecordOut'
                title: Response List Milk Records Api V1 Production Milk Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - production
      summary: Record Milk
      description: "Validation rule (tech spec \xA714): \"Milk from an animal under\n\
        withdrawal must show a hard warning or be blocked from sale\ndestination.\"\
        \ This endpoint hard-blocks it, matching RULE-WITHDRAWAL\nin the recommendation\
        \ engine."
      operationId: record_milk_api_v1_production_milk_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MilkRecordCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Record Milk Api V1 Production Milk Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/production/eggs:
    get:
      tags:
      - production
      summary: List Egg Records
      operationId: list_egg_records_api_v1_production_eggs_get
      security:
      - HTTPBearer: []
      parameters:
      - name: farm_id
        in: query
        required: true
        schema:
          type: string
          title: Farm Id
      - name: days
        in: query
        required: false
        schema:
          type: integer
          default: 30
          title: Days
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EggRecordOut'
                title: Response List Egg Records Api V1 Production Eggs Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - production
      summary: Record Eggs
      description: "Validation rule (tech spec \xA714): \"Sellable + broken + consumed\
        \ +\nhatched + wasted cannot exceed total eggs.\""
      operationId: record_eggs_api_v1_production_eggs_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EggRecordCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Record Eggs Api V1 Production Eggs Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/production/harvest:
    get:
      tags:
      - production
      summary: List Harvest Records
      operationId: list_harvest_records_api_v1_production_harvest_get
      security:
      - HTTPBearer: []
      parameters:
      - name: farm_id
        in: query
        required: true
        schema:
          type: string
          title: Farm Id
      - name: days
        in: query
        required: false
        schema:
          type: integer
          default: 90
          title: Days
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HarvestRecordOut'
                title: Response List Harvest Records Api V1 Production Harvest Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - production
      summary: Record Harvest
      operationId: record_harvest_api_v1_production_harvest_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HarvestRecordCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Record Harvest Api V1 Production Harvest Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    DailyHarvestCreate:
      properties:
        field_id:
          type: string
          title: Field Id
        planting_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Planting Id
        crop_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Crop Id
        product_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Product Name
        total_quantity:
          type: number
          title: Total Quantity
        sellable_quantity:
          anyOf:
          - type: number
          - type: 'null'
          title: Sellable Quantity
        waste_quantity:
          type: number
          title: Waste Quantity
          default: 0
        unit:
          type: string
          title: Unit
          default: kg
        destination:
          anyOf:
          - type: string
          - type: 'null'
          title: Destination
        recorded_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Recorded At
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      required:
      - field_id
      - total_quantity
      title: DailyHarvestCreate
      description: "Tech spec \xA717 \"How much was harvested today?\" \u2014 the\
        \ sellable/waste\nsplit is the point: it is what makes waste and cost-per-kg\
        \ real\nnumbers instead of estimates."
    DailyHarvestOut:
      properties:
        id:
          type: string
          title: Id
        field_id:
          type: string
          title: Field Id
        product_name:
          type: string
          title: Product Name
        total_quantity:
          type: number
          title: Total Quantity
        sellable_quantity:
          type: number
          title: Sellable Quantity
        waste_quantity:
          type: number
          title: Waste Quantity
        unit:
          type: string
          title: Unit
        recorded_at:
          type: string
          format: date-time
          title: Recorded At
        inventory_item_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Inventory Item Id
        inventory_qty_after:
          anyOf:
          - type: number
          - type: 'null'
          title: Inventory Qty After
      type: object
      required:
      - id
      - field_id
      - product_name
      - total_quantity
      - sellable_quantity
      - waste_quantity
      - unit
      - recorded_at
      title: DailyHarvestOut
    EggRecordCreate:
      properties:
        flock_id:
          type: string
          title: Flock Id
        total_eggs:
          type: integer
          title: Total Eggs
        sellable_eggs:
          type: integer
          title: Sellable Eggs
          default: 0
        broken_eggs:
          type: integer
          title: Broken Eggs
          default: 0
        consumed:
          type: integer
          title: Consumed
          default: 0
        hatched:
          type: integer
          title: Hatched
          default: 0
        wasted:
          type: integer
          title: Wasted
          default: 0
        recorded_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Recorded At
      type: object
      required:
      - flock_id
      - total_eggs
      title: EggRecordCreate
      description: "POST /production/eggs.\n\nValidation rule (tech spec \xA714):\
        \ \"Sellable + broken + consumed +\nhatched + wasted cannot exceed total eggs.\""
    EggRecordOut:
      properties:
        id:
          type: string
          title: Id
        flock_id:
          type: string
          title: Flock Id
        total_eggs:
          type: integer
          title: Total Eggs
        sellable_eggs:
          type: integer
          title: Sellable Eggs
        broken_eggs:
          type: integer
          title: Broken Eggs
        consumed:
          type: integer
          title: Consumed
        hatched:
          type: integer
          title: Hatched
        wasted:
          type: integer
          title: Wasted
        recorded_at:
          type: string
          format: date-time
          title: Recorded At
      type: object
      required:
      - id
      - flock_id
      - total_eggs
      - sellable_eggs
      - broken_eggs
      - consumed
      - hatched
      - wasted
      - recorded_at
      title: EggRecordOut
...
    FieldCreate:
      properties:
        name:
          type: string
          title: Name
        field_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Field Code
        area_value:
          anyOf:
          - type: number
          - type: 'null'
          title: Area Value
        area_unit:
          anyOf:
          - type: string
          - type: 'null'
          title: Area Unit
          default: m2
        location_label:
          anyOf:
          - type: string
          - type: 'null'
          title: Location Label
        soil_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Soil Type
        irrigation_method:
          anyOf:
          - type: string
          - type: 'null'
          title: Irrigation Method
        status:
          type: string
          title: Status
          default: active
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      required:
      - name
      title: FieldCreate
    FieldOut:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        crop_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Crop Type
        area_value:
          anyOf:
          - type: number
          - type: 'null'
          title: Area Value
        area_unit:
          anyOf:
          - type: string
          - type: 'null'
          title: Area Unit
        stage:
          anyOf:
          - type: string
          - type: 'null'
          title: Stage
        expected_harvest_date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expected Harvest Date
        est_yield_kg:
          anyOf:
          - type: number
          - type: 'null'
          title: Est Yield Kg
        field_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Field Code
        location_label:
          anyOf:
          - type: string
          - type: 'null'
          title: Location Label
        soil_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Soil Type
        irrigation_method:
          anyOf:
          - type: string
          - type: 'null'
          title: Irrigation Method
        status:
          type: string
          title: Status
          default: active
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      required:
      - id
      - name
      title: FieldOut
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    HarvestRecordOut:
      properties:
        id:
          type: string
          title: Id
        field_id:
          type: string
          title: Field Id
        product_name:
          type: string
          title: Product Name
        quantity:
          type: number
          title: Quantity
        unit:
          type: string
          title: Unit
        waste_qty:
          type: number
          title: Waste Qty
        destination:
          anyOf:
          - type: string
          - type: 'null'
          title: Destination
        recorded_at:
          type: string
          format: date-time
          title: Recorded At
      type: object
      required:
      - id
      - field_id
      - product_name
      - quantity
      - unit
      - waste_qty
      - recorded_at
      title: HarvestRecordOut
...
    MilkRecordCreate:
      properties:
        animal_id:
          type: string
          title: Animal Id
        session:
          type: string
          title: Session
        liters:
          type: number
          title: Liters
        destination:
          type: string
          title: Destination
          default: stored
        quality_status:
          type: string
          title: Quality Status
          default: normal
        recorded_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Recorded At
        recorded_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Recorded By
      type: object
      required:
      - animal_id
      - session
      - liters
      title: MilkRecordCreate
      description: "POST /production/milk.\n\nValidation rule (tech spec \xA714):\
        \ \"Liters >= 0; destination sale blocked\nor hard-warned if withdrawal active.\"\
        \ The hard block (not just a\nwarning) happens in services/production.py,\
        \ which has access to the\nanimal's current withdrawal_until date."
    MilkRecordOut:
      properties:
        id:
          type: string
          title: Id
        animal_id:
          type: string
          title: Animal Id
        session:
          type: string
          title: Session
        liters:
          type: number
          title: Liters
        quality_status:
          type: string
          title: Quality Status
        destination:
          type: string
          title: Destination
        recorded_at:
          type: string
          format: date-time
          title: Recorded At
        recorded_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Recorded By
      type: object
      required:
      - id
      - animal_id
      - session
      - liters
      - quality_status
      - destination
      - recorded_at
      title: MilkRecordOut
...
    ObservationCreate:
      properties:
        farm_id:
          type: string
          title: Farm Id
        entity_type:
          type: string
          title: Entity Type
        entity_id:
          type: string
          title: Entity Id
        observation_type:
          type: string
          title: Observation Type
        quality:
          type: string
          title: Quality
          default: human_observed
        value_numeric:
          anyOf:
          - type: number
          - type: 'null'
          title: Value Numeric
        value_text:
          anyOf:
          - type: string
          - type: 'null'
          title: Value Text
        unit:
          anyOf:
          - type: string
          - type: 'null'
          title: Unit
        severity:
          anyOf:
          - type: string
          - type: 'null'
          title: Severity
        observed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Observed At
        observer_id:
          type: string
          title: Observer Id
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      required:
      - farm_id
      - entity_type
      - entity_id
      - observation_type
      - observer_id
      title: ObservationCreate
      description: "POST /observations.\n\nConstitution: \"Workers record observations.\
        \ Workers do not diagnose.\"\nThere is deliberately no `diagnosis` field here\
        \ \u2014 see\nschemas/health.TreatmentCreate for the manager/vet-gated path."
    ObservationOut:
      properties:
        id:
          type: string
          title: Id
        farm_id:
          type: string
          title: Farm Id
        entity_type:
          type: string
          title: Entity Type
        entity_id:
          type: string
          title: Entity Id
        observation_type:
          type: string
          title: Observation Type
        quality:
          type: string
          title: Quality
        confidence:
          type: number
          title: Confidence
        value_numeric:
          anyOf:
          - type: number
          - type: 'null'
          title: Value Numeric
        value_text:
          anyOf:
          - type: string
          - type: 'null'
          title: Value Text
        unit:
          anyOf:
          - type: string
          - type: 'null'
          title: Unit
        severity:
          anyOf:
          - type: string
          - type: 'null'
          title: Severity
        observed_at:
          type: string
          format: date-time
          title: Observed At
        observer_id:
          type: string
          title: Observer Id
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      required:
      - id
      - farm_id
      - entity_type
      - entity_id
      - observation_type
      - quality
      - confidence
      - value_numeric
      - value_text
      - unit
      - severity
      - observed_at
      - observer_id
      - notes
      title: ObservationOut
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer